Change query_debounce to prevent freezing
On lower end machines 0.1s is not enough time for the OS to process the keyboard input and display results before another input is sent.  Thus you get a long pause before results appear.  I noticed this awhile back and only am just now getting around to suggesting a fix.  I tested this change locally, and it seemed to resolve the issue.